Rename package che_rest -> server
